no longer build sysupgrade image
Using the image build to generate a sysupgrade image was a work-around
to allow normal users to use snapshot images which would include LuCI.
Now that an official release supporting the device is about to be
available, let's just use that instead.

This implicitely changes back to 'wolfssl' and also no longer
pre-installs 'auc', 'luci-app-attendedsysupgrade' and a few other
packages users may easily add again if they want to.