Pin openwrt release to 24.10.0-rc5 to work around APK/OPKG incompatibility
See #204. The latest snapshots of OpenWrt have transitioned to using APK
instead of OPKG for package management, and this installer relies on
OPKG. So the quick workaround is to explicitly pin the latest version
of OpenWrt that still used OPKG.

Eventually, it will make sense to port this installer to APK, but at
the moment there are problems with post-install scripts of some packages
that we depend on, namely https://github.com/openwrt/openwrt/issues/17439

So for the time being, the best strategy to keep this repo working is
to just pin the version.